// Qanat.Web/src/app/shared/generated/model/zone-display-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ZoneDisplayDto { 
    ZoneID?: number;
    ZoneGroupID?: number;
    ZoneName?: string | null;
    ZoneColor?: string | null;
    ZoneAccentColor?: string | null;
    ZoneGroupDisplayToAccountHolders?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ZoneDisplayDtoForm { 
    ZoneID?: FormControl<number>;
    ZoneGroupID?: FormControl<number>;
    ZoneName?: FormControl<string>;
    ZoneColor?: FormControl<string>;
    ZoneAccentColor?: FormControl<string>;
    ZoneGroupDisplayToAccountHolders?: FormControl<boolean>;
}

export class ZoneDisplayDtoFormControls { 
    public static ZoneID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ZoneGroupID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ZoneName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ZoneColor = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ZoneAccentColor = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ZoneGroupDisplayToAccountHolders = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}